Added video preview generation
Added start and duration to startCompress method along with
includeAudio and frame rate parameters. This is made in order
to create smaller videos that can be used for previews.
Currently on Android all parameters are used, on iOS it's not
possible with current library to set frame rate, and removing
audio from video is relativly complicated, so for the initial
implementation it is left out of iOS impelementation.